// src/pages/Products/utils/types/TFurniture.ts
export type TFurniture = {
  _id: string;
  name: string;
  productId: string;
  quantity: number;
  price: number;
  type: string;
  dimensions: string;
  image: string;
  category: string;
  warranty: string;
  description: string;
  color: string;
  material: string;
  createdBy: string;
  createdAt: string;
  updatedAt: string;
};